backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
PyJWT==2.8.0
pymongo==4.6.0
